+++ examples/custom_checkers/e9992_top_level_code.py
@@ -1,0 +1,5 @@
+def example_function(name: str) -> str:
+    return f'Hello {name}!'
+
+
+print(example_function('Fred'))  # error on this line