+++ tests/test_contracts.py
@@ -16,12 +16,32 @@ def test_nullary_return_int() -> None:
     nullary()
 
 
+def test_nullary_return_float() -> None:
+    """Calling a nullary function with the correct return type (float)."""
+
+    @check_contracts
+    def nullary() -> float:
+        return 1.0
+
+    nullary()
+
+
 def test_nullary_return_dict() -> None:
     """Calling a nullary function with the correct return type (Dict)."""
 
     @check_contracts
     def nullary() -> Dict[str, int]:
         return {"one": 1}
+
+    nullary()
+
+
+def test_nullary_return_bool() -> None:
+    """Calling a nullary function with the correct return type (Bool)."""
+
+    @check_contracts
+    def nullary() -> bool:
+        return True
 
     nullary()
 
@@ -58,6 +78,50 @@ def test_nullary_return_dict_wrong() -> None:
         nullary()
 
 
+def test_nullary_return_int_float_error() -> None:
+    """Calling a nullary function with return type int, got float"""
+
+    @check_contracts
+    def nullary() -> int:
+        return 1.5
+
+    with pytest.raises(AssertionError):
+        nullary()
+
+
+def test_nullary_return_float_int_error() -> None:
+    """Calling a nullary function with return type float, got int"""
+
+    @check_contracts
+    def nullary() -> float:
+        return 1
+
+    with pytest.raises(AssertionError):
+        nullary()
+
+
+def test_nullary_return_int_bool_error() -> None:
+    """Calling a nullary function with return type int, got bool"""
+
+    @check_contracts
+    def nullary() -> int:
+        return True
+
+    with pytest.raises(AssertionError):
+        nullary()
+
+
+def test_nullary_return_bool_int_error() -> None:
+    """Calling a nullary function with return type bool, got int"""
+
+    @check_contracts
+    def nullary() -> bool:
+        return 1
+
+    with pytest.raises(AssertionError):
+        nullary()
+
+
 def test_nullary_no_return_type() -> None:
     """Calling a nullary function with no specified return type passes."""
 
@@ -78,6 +142,12 @@ def test_my_sum_list_int_argument() -> None:
     _my_sum([1, 2, 3])
 
 
+def test_my_sum_float_argument() -> None:
+    """Calling _my_sum with a list of floats fails type check."""
+    with pytest.raises(AssertionError):
+        _my_sum([1.5, 2.0])
+
+
 def test_my_sum_empty_list_argument() -> None:
     """Calling _my_sum with an empty list passes type check."""
     _my_sum([])
@@ -87,6 +157,66 @@ def test_my_sum_list_mixed_argument() -> None:
     """Calling _my_sum with a list containing not just ints fails type check."""
     with pytest.raises(AssertionError):
         _my_sum([1, 2, "hello"])
+
+
+def test_parameter_int() -> None:
+    @check_contracts
+    def parameter_int(num: int) -> None:
+        return None
+
+    parameter_int(1)
+
+
+def test_parameter_float() -> None:
+    @check_contracts
+    def parameter_float(num: float) -> None:
+        return None
+
+    parameter_float(1.0)
+
+
+def test_parameter_bool() -> None:
+    @check_contracts
+    def parameter_bool(result: bool) -> None:
+        return None
+
+    parameter_bool(True)
+
+
+def test_parameter_int_float_error() -> None:
+    @check_contracts
+    def parameter_int(num: int) -> None:
+        return None
+
+    with pytest.raises(AssertionError):
+        parameter_int(1.0)
+
+
+def test_parameter_float_int_error() -> None:
+    @check_contracts
+    def parameter_float(num: float) -> None:
+        return None
+
+    with pytest.raises(AssertionError):
+        parameter_float(1)
+
+
+def test_parameter_int_bool_error() -> None:
+    @check_contracts
+    def parameter_int(num: int) -> None:
+        return None
+
+    with pytest.raises(AssertionError):
+        parameter_int(True)
+
+
+def test_parameter_bool_int_error() -> None:
+    @check_contracts
+    def parameter_bool(result: bool) -> None:
+        return None
+
+    with pytest.raises(AssertionError):
+        parameter_bool(1)
 
 
 @check_contracts